new helm action gnus-notes-insert-quick-note
This new function inserts a quick list item with the article (email) timestamp and a
org-gnus link. Very convenient to write some notes afterwards. The actual note is
created in also new function gnus-notes-quick-note.

changed the gnus-notes-format-time-string to not display seconds

To conform closer to the org time format. now after the date info, only hours and
minutes are displayed. This makes the date info slightly more compact.